# lib/api-zod/src/generated/types/betaTesterSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BetaTesterSubmission {
  /** @minLength 1 */
  name: string;
  email: string;
  products?: string[];
  expertise?: string;
}