# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  /** @minLength 1 */
  name: string;
  email: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}