# lib/api-zod/src/generated/types/developerApplicationSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeveloperApplicationSubmission {
  /** @minLength 1 */
  name: string;
  email: string;
  url: string;
  expertise?: string;
  /** @minLength 1 */
  about: string;
  available?: boolean;
}